+++ lib/td_cache/redix/pool.ex
@@ -13,7 +13,8 @@ defmodule TdCache.Redix.Pool do
     # Specs for the Redix connections.
     children =
       for i <- 0..(@pool_size - 1) do
-        Supervisor.child_spec({Redix, host: redis_host, port: port, password: password, name: :"redix_#{i}"},
+        Supervisor.child_spec(
+          {Redix, host: redis_host, port: port, password: password, name: :"redix_#{i}"},
           id: {Redix, i}
         )
       end